changed time from int to long

+++ src/magic/model/MagicCardList.java
@@ -71,7 +71,7 @@ public class MagicCardList extends ArrayList<MagicCard> {
 		return index;
 	}
 	
-	public MagicCard getCard(final int id) {
+	public MagicCard getCard(final long id) {
 		for (final MagicCard card : this) {
 			if (card.getId()==id) {
 				return card;
